% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBFM-package.R
\docType{package}
\name{CBFM-package}
\alias{CBFM-package}
\title{CBFM: Community-Level Basis Function Modeling (CBFM)}
\description{
Spatio-temporal joint species distribution modeling via community-level basis functions. CBFMs are built on the idea of fixed-rank kriging (FRK), where spatial and/or temporally indexed basis functions are included to account for spatio-temporal correlations within and between species. The approach provides an alternative to the popular and powerful latent variable approach for joint species distribution modeling in community ecology, with the main advantage being that CBFMs can be computationally more efficient and scalable to fit. This is because, by using a basis function approach, CBFMs can be set up as a type of generalized additive model (GAM), from which we can leverage from existing methods developed for GAMs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fhui28/CBFM}
  \item Report bugs at \url{https://github.com/fhui28/CBFM/issues}
}

}
\author{
\strong{Maintainer}: Francis K.C. Hui \email{fhui28@gmail.com}

Authors:
\itemize{
  \item Francis K.C. Hui \email{fhui28@gmail.com}
  \item Chris Haak
}

}
\keyword{internal}
