% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.CBFM.R
\name{logLik.CBFM}
\alias{logLik.CBFM}
\title{Log-likelihood of a CBFM fit}
\usage{
\method{logLik}{CBFM}(object, ...)
}
\arguments{
\item{object}{An object of class "CBFM".}

\item{...}{Not used in this case.}
}
\value{
Returns an object of class "logLik".  This is a number with at least one attribute, "df" (degrees of freedom), giving the number of (estimated) parameters in the CBFM model; please see see \code{\link[stats:logLik]{stats::logLik()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extracts the log-likelihood from a fitted \code{CBFM} object.
}
\details{
In this package, CBFM are fitted using maximized penalized quasi-likelihood or PQL estimation. In turn, this function returns the maximized PQL value of the \code{CBFM} object at convergence. Note that accordingly, the degrees of freedom calculation is based on treating the regression coefficients related to the basis functions as fixed effects, meaning the calculation only involves the number of regression coefficients in the model plus any nusiance parameters (see Hui et al., 2017, Hui, 2021). The community-level covariance matrices are not involved in the calculation of the degrees of freedom.

Currently the idea of \emph{effective} degrees of freedom have not yet been implemented in this package.
}
\examples{
\donttest{
library(autoFRK)
library(FRK)
library(MASS)
library(mvabund)
library(mvtnorm)
library(ROCR)
library(sp)
library(RandomFields)
library(tidyverse)

##------------------------------
## Example 1: Fitting a CBFM to spatial multivariate presence-absence data 
## simulated from a spatial latent variable model
## Please note the data generation process (thus) differs from CBFM.
##------------------------------
set.seed(2021)
num_sites <- 500 # 500 (units) sites 
num_spp <- 50 # Number of species
num_X <- 4 # Number of regression slopes

spp_slopes <- matrix(runif(num_spp * num_X, -1, 1), nrow = num_spp)
spp_intercepts <- runif(num_spp, -2, 0)

# Simulate spatial coordinates and environmental covariate components
xy <- data.frame(x = runif(num_sites, 0, 5), y = runif(num_sites, 0, 5))
X <- rmvnorm(num_sites, mean = rep(0,4)) 
colnames(X) <- c("temp", "depth", "chla", "O2")
dat <- data.frame(xy, X)
mm <- model.matrix(~ temp + depth + chla + O2 - 1, data = dat) \%>\% 
scale \%>\% 
as.matrix

# Simulate latent variable component
true_lvs <- RFsimulate(model = RMexp(var=1, scale=2), 
x = xy$x, y = xy$y, n = 2)@data \%>\% 
as.matrix
spp_loadings <- matrix(runif(num_spp * 2, -1, 1), nrow = num_spp) 
set.seed(NULL)

# Simulate spatial multivariate abundance data (presence-absence)
eta <- tcrossprod(cbind(1,mm), cbind(spp_intercepts,spp_slopes)) + 
tcrossprod(true_lvs, spp_loadings)
simy <- matrix(rbinom(num_sites * num_spp, size = 1, 
prob = binomial()$linkinv(eta)), nrow = num_sites)
rm(X, mm, spp_loadings, true_lvs, xy, eta)


# Set up spatial basis functions for CBFM -- Most practitioners will start here! 
num_basisfunctions <- 25 # Number of spatial basis functions to use
basisfunctions <- mrts(dat[,c("x","y")], num_basisfunctions) \%>\% 
as.matrix \%>\%
{.[,-(1)]} # Remove the first intercept column

# Fit CBFM 
useformula <- ~ temp + depth + chla + O2
fitcbfm <- CBFM(y = simy, formula_X = useformula, data = dat, 
B_space = basisfunctions, family = binomial(), control = list(trace = 1))

logLik(fitcbfm)
}

}
\references{
Hui, F. K. C., Mueller, S., and Welsh, A. H. (2017). Joint selection in mixed models using regularized PQL. Journal of the American Statistical Association, 112, 1323-1333.

Hui, F. K. C.(2021). On the use of a penalized quasilikelihood information criterion for generalized linear mixed models. Biometrika, 108, 353-365.
}
\seealso{
\code{\link[=CBFM]{CBFM()}} for fitting CBFMs, and \code{\link[=AIC.CBFM]{AIC.CBFM()}} and \code{\link[=AICc.CBFM]{AICc.CBFM()}} for calculation various information criteria from a CBFM fit.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
