% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinate.CBFM.R
\name{ordinate.CBFM}
\alias{ordinate.CBFM}
\alias{ordinate}
\title{Construct spatio-temporal ordinations from a CBFM fit}
\usage{
\method{ordinate}{CBFM}(
  object,
  num_comp = 2,
  new_B_space = NULL,
  new_B_time = NULL,
  new_B_spacetime = NULL,
  alpha = 0.5,
  ...
)
}
\arguments{
\item{object}{An object of class \code{CBFM}.}

\item{num_comp}{The number of ordination axes to construct.}

\item{new_B_space}{A matrix of new spatial basis functions at which ordinations are to be constructed. If this is not provided, then ordinations based on the original set of basis functions are returned. If at least one of \code{new_B_space/new_B_time/new_B_spacetime} are provided, then it is \emph{assumed} ordinations at new basis function values are desired.}

\item{new_B_time}{A matrix of new temporal basis functions at which ordinations are to be constructed. If this is not provided, then ordinations based on the original set of basis functions are returned. If at least one of \code{new_B_space/new_B_time/new_B_spacetime} are provided, then it is \emph{assumed} ordinations at new basis function values are desired.}

\item{new_B_spacetime}{A matrix of new spatio-temporal basis functions at which ordinations are to be constructed. If this is not provided, then ordinations based on the original set of basis functions are returned. If at least one of \code{new_B_space/new_B_time/new_B_spacetime} are provided, then it is \emph{assumed} ordinations at new basis function values are desired.}

\item{alpha}{A numeric scalar between 0 and 1 that controls the relative scaling of the score and corresponding loadings, when constructing a biplot. Defaults to 0.5, which gives equal weighting to above.}

\item{...}{Not used.}
}
\value{
A list with the following items:
\item{scores: }{A matrix of observational unit-specific scores.}
\item{loadings: }{A matrix of species-specific loadings.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates observational unit-specific scores and corresponding species-specific loadings from a fitted \code{CBFM} object, potentially at new values of the basis functions. which in turn can be used as a means of spatio-temporal ordination.
}
\details{
For a CBFM, one way of performing spatio-temporal ordination is as follows. For observational unit \eqn{i=1,\ldots,N} and species \eqn{j=1,\ldots,m}, recall that for the purposes of this package, the CBFM is defined by the mean model:

\deqn{g(\mu_{ij}) = \eta_{ij} = x_i^\top\beta_j + b_i^\top a_j,}

where \eqn{g(.)} is a known link function, \eqn{x_i} denotes a vector of predictors for unit \eqn{i} i.e., the \eqn{i}-th row from the created model matrix, \eqn{\beta_j} denotes the corresponding regression coefficients for species \eqn{j}, \eqn{b_i} denotes a vector of spatial and/or temporal basis functions for unit \eqn{i} , and \eqn{a_j} denotes the corresponding regression coefficients for species \eqn{j}.

Consider now the estimates values of \eqn{b_i^\top a_j} and construct an \eqn{N \times m} matrix of estimated linear predictors with this. We can then construct an ordination from this matrix, albeit in a bit of an ad-hoc manner, by performing a singular value decomposition to obtain matrices of left and and right singular vectors, along with a diagonal matrix of singular values in descending order. The left singular vectors correspond to observational unit-specific scores (similar to latent variables from an latent variable model; see for example Hui et al., 2015, Warton et al., 2015, Niku et al., 2019, van der Veen, 2021) while the right singular vectors correspond to the corresponding species-specific loadings (similar to latent variables from an latent variable model). By taking only the first \code{num_comp} of the left and right singular vectors and scaling according to the singular values, we can then use these as a means of spatial and/or temporal ordination.

The scores and loadings and thus ordinations can also be constructed at a new set of observational units. This is achieved by supplying the values of \code{new_B_space/new_B_time/new_B_spacetime} as appropriate, similar to how \code{\link[=predict.CBFM]{predict.CBFM()}} functions. Ordinations at new observational units are often desired for spatial temporal ordinations e.g., if the ordination is to be constructed and visualized on a grid of spatial locations, and potentially over time.

Please note the current implementation does \emph{not} produce plots automatically: it only returns the scores and loadings, which the user then has to manually use to construct plots. Future versions of the package may change this!

#Warning
For zero-truncated count CBFMs, observational unit-specific scores are produced for all observational units, although perhaps these should be treated cautiously if the response matrix itself contains zero counts.
}
\examples{
\dontrun{
library(autoFRK)
library(FRK)
library(MASS)
library(mvabund)
library(mvtnorm)
library(ROCR)
library(sp)
library(geoR)
library(tidyverse)

##------------------------------
## **Example 1: Fitting a CBFM to spatial multivariate presence-absence data**
## simulated from a spatial latent variable model
## Please note the data generation process (thus) differs from CBFM.
##------------------------------
set.seed(2021)
num_sites <- 500 # 500 (units) sites 
num_spp <- 50 # Number of species
num_X <- 4 # Number of regression slopes

spp_slopes <- matrix(runif(num_spp * num_X, -1, 1), nrow = num_spp)
spp_intercepts <- runif(num_spp, -2, 0)

# Simulate spatial coordinates and environmental covariate components
xy <- data.frame(x = runif(num_sites, 0, 5), y = runif(num_sites, 0, 5))
X <- rmvnorm(num_sites, mean = rep(0,4)) 
colnames(X) <- c("temp", "depth", "chla", "O2")
dat <- data.frame(xy, X)
mm <- model.matrix(~ temp + depth + chla + O2 - 1, data = dat) \%>\% 
scale \%>\% 
as.matrix

# Simulate latent variable component
true_lvs <- grf(grid = cbind(xy$x, xy$y), nsim = 2, cov.model = "exponential",
cov.pars = c(1, 2))$data \%>\%
     as.matrix
spp_loadings <- matrix(runif(num_spp * 2, -1, 1), nrow = num_spp)
set.seed(NULL)

# Simulate spatial multivariate abundance data (presence-absence)
eta <- tcrossprod(cbind(1,mm), cbind(spp_intercepts,spp_slopes)) + 
tcrossprod(true_lvs, spp_loadings)
simy <- matrix(rbinom(num_sites * num_spp, size = 1, 
prob = binomial()$linkinv(eta)), nrow = num_sites)
rm(X, mm, spp_loadings, true_lvs, xy, eta)


# Set up spatial basis functions for CBFM -- Most users will start here! 
num_basisfunctions <- 25 # Number of spatial basis functions to use
basisfunctions <- mrts(dat[,c("x","y")], num_basisfunctions) \%>\% 
as.matrix \%>\%
{.[,-(1)]} # Remove the first intercept column

# Fit CBFM 
useformula <- ~ temp + depth + chla + O2
fitcbfm <- CBFM(y = simy, formula = useformula, data = dat, 
B_space = basisfunctions, family = binomial(), control = list(trace = 1))

getords <- ordinate(fitcbfm)

orddat <- data.frame(dat[,c("x","y")], getords$scores) 

ggplot(orddat, aes(x = x, y = y, color = Axis1)) +
geom_point() +
scale_color_viridis_c() +
labs(x = "x coordinate", y = "y coordinate", color = "Site score 1") +
theme_bw()

ggplot(orddat, aes(x = x, y = y, color = Axis2)) +
geom_point() +
scale_color_viridis_c() +
labs(x = "x coordinate", y = "y coordinate", color = "Site score 2") +
theme_bw()
}

}
\references{
Hui, F. K. C., Taskinen, S., Pledger, S., Foster, S. D., and Warton, D. I. (2015). Model‐based approaches to unconstrained ordination. Methods in Ecology and Evolution, 6, 399-411.

Niku, J., Hui, F. K. C., Taskinen, S., and Warton, D. I. (2019). gllvm: Fast analysis of multivariate abundance data with generalized linear latent variable models in R. Methods in Ecology and Evolution, 10, 2173-2182.

van der Veen, B., Hui, F. K. C., Hovstad, K. A., Solbu, E. B., & O'Hara, R. B. (2021). Model‐based ordination for species with unequal niche widths. Methods in Ecology and Evolution. In press.

Warton, D. I., Blanchet, F. G., O'Hara, R. B., Ovaskainen, O., Taskinen, S., Walker, S. C., and Hui, F. K. C. (2015). So many variables: joint modeling in community ecology. Trends in Ecology and Evolution, 30, 766-779.
}
\seealso{
\code{\link[=CBFM]{CBFM()}} for fitting CBFMs, \code{\link[=corB]{corB()}} for calculating residual between-species (cross-)correlations due to the basis functions, \code{\link[=predict.CBFM]{predict.CBFM()}} for constructing predictions from a CBFM fit.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
