% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CBFM.R
\name{predict.CBFM}
\alias{predict.CBFM}
\title{Construct predictions for a CBFM fit}
\usage{
\method{predict}{CBFM}(
  object,
  newdata = NULL,
  manualX = NULL,
  new_B_space = NULL,
  new_B_time = NULL,
  new_B_spacetime = NULL,
  type = "link",
  se_fit = FALSE,
  coverage = 0.95,
  ncores = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class "CBFM".}

\item{newdata}{A data frame containing the values of the covariates at which predictions are to be calculated. If this is not provided, then predictions corresponding to the original data are returned. If \code{newdata} is provided then it should contain all the variables needed for prediction, that is, it can construct a model matrix from this as \code{object$formula_X}.}

\item{manualX}{A manually supplied model matrix at which predictions are to be calculated. This can be used if for some reason the user wants to supply a very custom model matrix for calculating predictions. Note supply of this overrides any supplied \code{newdata} argument. The number of columns in \code{manualX} should equal to \code{ncol(object$betas)}.}

\item{new_B_space}{A matrix of new spatial basis functions at which predictions are to be calculated. If this is not provided, then predictions corresponding to the original \code{B_space} argument are returned. Please note this should only be supplied if \code{B_space} was supplied in the original CBFM fit.}

\item{new_B_time}{A matrix of new temporal basis functions at which predictions are to be calculated. If this is not provided, then predictions corresponding to the original \code{B_time} argument are returned. Please note this should only be supplied if \code{B_time} was supplied in the original CBFM fit.}

\item{new_B_spacetime}{A matrix of new spatio-temporal basis functions at which predictions are to be calculated. If this is not provided, then predictions corresponding to the original \code{B_spacetime} argument are returned. Please note this should only be supplied if \code{B_spacetime} was supplied in the original CBFM fit.}

\item{type}{The type of prediction required. The default \code{type = "link"} is on the scale of the linear predictors. Alternatively, \code{type = "response"} returns predictions on the scale of the response variable. For example, the predictions for a binomial CBFM are the predicted probabilities.
Note that zero-inflated distributions, \code{type = "link"} returns the predicted linear predictor of the non-zero-inflated component, consistent with what \code{object$linear_predictor} returns. But for \code{type = "response"}, it returns the \emph{actual predicted mean values} of the distribution, consistent with what \code{object$fitted} returns.
A third option is given by \code{type  = "lpmatrix"}, which returns the model matrix of the covariates constructing (potentially) using \code{newdata}. In \code{\link[mgcv:predict.gam]{mgcv::predict.gam()}}, this is also known as the linear predictor matrix. Note under this option, arguments such as \code{new_B_space/new_B_time/new_B_spacetime, se_fit} and \code{coverage} are irrelevant and hence ignored.}

\item{se_fit}{When this is set to \code{TRUE} (not default), then standard error estimates are returned for each predicted value.}

\item{coverage}{The coverage probability of the uncertainty intervals for prediction. Defaults to 0.95, which corresponds to 95\% uncertainty intervals.}

\item{ncores}{To speed up calculation of the standard error estimates, parallelization can be performed, in which case this argument can be used to supply the number of cores to use in the parallelization. Defaults to \code{detectCores()-1}.}

\item{...}{Not used.}
}
\value{
If \code{se_fit = TRUE}, then a list with the following components (if applicable) is returned:
\item{fit: }{A matrix of predicted values.}
\item{stderr: }{A matrix of standard errors associated with the uncertainty intervals for the predictions.}
\item{lower: }{A matrix of the lower bound of the uncertainty intervals for the predictions.}
\item{upper: }{A matrix of the upper bound of the uncertainty intervals for the predictions.}
Otherwise if \code{se_fit = FALSE}, then a matrix of predicted values is returned.

Other if \code{type = "lpmatrix"}, then a model matrix with the number of rows equal to either the number of rows if \code{object$y} (if \code{newdata} is not supplied) or the number of rows in \code{newdata} when it is supplied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a fitted \code{CBFM} object and produces predictions given (potentially) a new set of observational units with their corresponding covariate and basis function functions. Predictions can be accompanied by standard errors, based on the Bayesian covariance matrix of the parameter estimates. As another option, the function can return the model matrix of the covariates constructed (potentially) using at the new set of observational units; in \code{\link[mgcv:predict.gam]{mgcv::predict.gam()}} this is also known as the linear predictor matrix.
}
\details{
The standard errors produced by \code{predict.CBFM} are based on the Bayesian posterior covariance matrix of the estimated parameters from the fitted \code{CBFM} object, and associated uncertainty intervals are obtained based on the associated large sample normality result of i.e., basically a Gaussian approximation to the posterior to, the estimated parameters. This construction is similar to \code{\link[mgcv:predict.gam]{mgcv::predict.gam()}}.

The functions tries to avoid using simulation for constructing the uncertainty intervals for the predictions (to minimize computational burden) However, in some cases it will fallback to doing so when it (blame Francis!) can not find a simple way to construct these intervals e.g., for the zero-inflated Poisson distribution when \code{type = "response"}.
}
\examples{
\donttest{
# Please see examples in the help file for the main CBFM function 
}

}
\seealso{
\code{\link[=CBFM]{CBFM()}} for fitting CBFMs, \code{\link[=fitted.values.CBFM]{fitted.values.CBFM()}} for obtaining fitted values from a CBFM fit, and \code{\link[=residuals.CBFM]{residuals.CBFM()}} for calculating various types of residuals.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
