% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorproduct.R
\name{tensorproduct}
\alias{tensorproduct}
\title{Tensor product}
\usage{
tensorproduct(B1, B2)
}
\arguments{
\item{B1}{A matrix.}

\item{B2}{A matrix.}
}
\value{
A matrix resulting from the tensor product, with the number of columns equal to \code{ncol(B1)*ncol(B2)}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simple function used to calculate the tensor product of two matrices.
}
\details{
For the purposes of this package, this function may be used for constructing a matrix basis functions formed from the tensor product of two individual basis functions. A prime example of this is the construction of \code{B_spacetime} for use in the main function \code{\link[=CBFM]{CBFM()}}, where the spatio-temporal basis functions may be formed as a tensor product of spatial and temporal basis functions.

The tensor product is formed by multiplying the first column of \code{B2} with \code{B1}, then the second column of \code{B2} with \code{B1}, and so on, working through all the columns of \code{B2}.
}
\section{Warnings}{
No attempt is made by the function to made the resulting matrix sparse, unless the supplied \code{B1} and \code{B2} are themselves sparse. Therefore, please be careful of how much memory the resulting object uses!
}

\examples{
\dontrun{
# Please see Example 3 in the help file for the main CBFM function 
}

}
\seealso{
\code{\link[=CBFM]{CBFM()}} for fitting CBFMs, with one example using a tensor product.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
