% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipoisson.R
\name{zipoisson}
\alias{zipoisson}
\title{Family object for zero-inflated Poisson distribution}
\usage{
zipoisson()
}
\value{
An object of class "family".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A specified family object for the zero-inflated Poisson distribution, using the log link function for the Poisson distribution component.
}
\details{
This family object was created specifically for fitting CBFMs.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
