% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betalogitfam.R
\name{betalogitfam}
\alias{betalogitfam}
\title{Family object for beta distribution}
\usage{
betalogitfam()
}
\value{
An object of class "family".
}
\description{
\verb{r lifecycle::badge("stable")}

A specified family object for the beta distribution using the logit link function.
}
\details{
This family object was created specifically for fitting CBFMs.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
