% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb2.R
\name{nb2}
\alias{nb2}
\title{Family object for negative binomial distribution}
\usage{
nb2()
}
\value{
An object of class "family".
}
\description{
\verb{r lifecycle::badge("stable")}

A specified family object for the negative binomial distribution using the log link function and a quadratic mean-variance relationship.
}
\details{
This family object was cretated specifically for fitting CBFMs to overdispersed count data
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
