% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corB.R, R/predict.CBFM.R
\name{predict.CBFM}
\alias{predict.CBFM}
\alias{predict.CBFM_hurdle}
\title{Construct (cross-)correlations due to basis functions for a CBFM fit}
\usage{
\method{predict}{CBFM}(
  object,
  newdata = NULL,
  manualX = NULL,
  new_B_space = NULL,
  new_B_time = NULL,
  new_B_spacetime = NULL,
  type = "link",
  se_fit = FALSE,
  coverage = 0.95,
  ncores = NULL,
  num_sims = 500,
  return_internals = FALSE,
  ...
)

\method{predict}{CBFM_hurdle}(
  object,
  newdata_pa = NULL,
  manualX_pa = NULL,
  new_B_space_pa = NULL,
  new_B_time_pa = NULL,
  new_B_spacetime_pa = NULL,
  newdata_count = NULL,
  manualX_count = NULL,
  new_B_space_count = NULL,
  new_B_time_count = NULL,
  new_B_spacetime_count = NULL,
  se_fit = FALSE,
  coverage = 0.95,
  ncores = NULL,
  num_sims = 500,
  ...
)

\method{predict}{CBFM}(
  object,
  newdata = NULL,
  manualX = NULL,
  new_B_space = NULL,
  new_B_time = NULL,
  new_B_spacetime = NULL,
  type = "link",
  se_fit = FALSE,
  coverage = 0.95,
  ncores = NULL,
  num_sims = 500,
  return_internals = FALSE,
  ...
)

\method{predict}{CBFM_hurdle}(
  object,
  newdata_pa = NULL,
  manualX_pa = NULL,
  new_B_space_pa = NULL,
  new_B_time_pa = NULL,
  new_B_spacetime_pa = NULL,
  newdata_count = NULL,
  manualX_count = NULL,
  new_B_space_count = NULL,
  new_B_time_count = NULL,
  new_B_spacetime_count = NULL,
  se_fit = FALSE,
  coverage = 0.95,
  ncores = NULL,
  num_sims = 500,
  ...
)
}
\arguments{
\item{object}{An object of class \code{CBFM} or \code{CBFM_hurdle}.}

\item{newdata}{A data frame containing the values of the covariates at which predictions are to be calculated. If this is not provided, then predictions corresponding to the original data are returned. If \code{newdata} is provided then it should contain all the variables needed for prediction, that is, it can construct a model matrix from this as \code{object$formula_X}.}

\item{manualX}{A manually supplied model matrix at which predictions are to be calculated. This can be used if for some reason the user wants to supply a very custom model matrix for calculating predictions. Note supply of this overrides any supplied \code{newdata} argument. The number of columns in \code{manualX} should equal to \code{ncol(object$betas)}.}

\item{new_B_space}{A matrix of new spatial basis functions at which predictions are to be calculated. If this is not provided, then predictions corresponding to the original \code{B_space} argument are returned. Please note this should only be supplied if \code{B_space} was supplied in the original CBFM fit.}

\item{new_B_time}{A matrix of new temporal basis functions at which predictions are to be calculated. If this is not provided, then predictions corresponding to the original \code{B_time} argument are returned. Please note this should only be supplied if \code{B_time} was supplied in the original CBFM fit.}

\item{new_B_spacetime}{A matrix of new spatio-temporal basis functions at which predictions are to be calculated. If this is not provided, then predictions corresponding to the original \code{B_spacetime} argument are returned. Please note this should only be supplied if \code{B_spacetime} was supplied in the original CBFM fit.}

\item{type}{The type of prediction required. The default \code{type = "link"} is on the scale of the linear predictors. Alternatively, \code{type = "response"} returns predictions on the scale of the response variable. For example, the predicted response for a binomial CBFM are the predicted probabilities.
Note that zero-inflated distributions, \code{type = "link"} returns the predicted linear predictor of the non-zero-inflated component, consistent with what \code{object$linear_predictors} returns. But \code{type = "response"} returns the \emph{actual predicted mean values} of the distribution, consistent with what \code{object$fitted} returns.
Similarly, for zero-truncated distributions, code{type = "link"} returns the predicted linear predictor of the base count distribution, consistent with what \code{object$linear_predictors} returns. But \code{type = "response"} returns the \emph{actual predicted mean values} of the zero-truncated distribution, consistent with what \code{object$fitted} returns.
A third option is given by \code{type  = "lpmatrix"}, which returns the model matrix of the covariates constructing (potentially) using \code{newdata}. In \code{\link[mgcv:predict.gam]{mgcv::predict.gam()}}, this is also known as the linear predictor matrix. Note under this option, arguments such as \code{new_B_space/new_B_time/new_B_spacetime/se_fit/coverage} are irrelevant and hence ignored.
Finally, for hurdle CBFMS, this option is not available as by default only predicted mean values i.e., \code{type = "response"} are possible.}

\item{se_fit}{When this is set to \code{TRUE} (not default), then standard error estimates are returned for each predicted value.}

\item{coverage}{The coverage probability of the uncertainty intervals for prediction. Defaults to 0.95, which corresponds to 95\% uncertainty intervals.}

\item{ncores}{To speed up calculation of the standard error estimates, parallelization can be performed, in which case this argument can be used to supply the number of cores to use in the parallelization. Defaults to \code{detectCores()-1}.}

\item{num_sims}{If simulation is required for constructing uncertainty intervals, then this specifies the number of Monte-Carlo examples to simulate.}

\item{return_internals}{Not used. Please ignore this argument!}

\item{...}{Not used.}

\item{newdata_pa}{For hurdle CBFM models, similar to \code{newdata} above but applied to the presence-absence component model.}

\item{manualX_pa}{For hurdle CBFM models, similar to \code{manual_X} above but applied to the presence-absence component model.}

\item{new_B_space_pa}{For hurdle CBFM models, similar to \code{new_B_space} above but applied to the presence-absence component model.}

\item{new_B_time_pa}{For hurdle CBFM models, similar to \code{new_B_time} above but applied to the presence-absence component model.}

\item{new_B_spacetime_pa}{For hurdle CBFM models, similar to \code{new_B_spacetime} above but applied to the presence-absence component model.}

\item{newdata_count}{For hurdle CBFM models, similar to \code{newdata} above but applied to the zero-truncated count component model.}

\item{manualX_count}{For hurdle CBFM models, similar to \code{manual_X} above but applied to the zero-truncated count component model.}

\item{new_B_space_count}{For hurdle CBFM models, similar to \code{new_B_space} above but applied to the zero-truncated count component model.}

\item{new_B_time_count}{For hurdle CBFM models, similar to \code{new_B_time} above but applied to the zero-truncated count component model.}

\item{new_B_spacetime_count}{For hurdle CBFM models, similar to \code{new_B_spacetime} above but applied to the zero-truncated count component model.}

\item{new_B_space2}{A second new matrix of new spatial basis functions at which cross-correlations are to be calculated. If this is supplied, then \code{new_B_space} must also be supplied, as the function assumes then the user desires calculation of cross-correlations due to these spatial basis functions.}

\item{new_B_time2}{A second new matrix of temporal basis functions at which cross-calculations are to be calculated.  If this is supplied, then \code{new_B_time} must also be supplied, as the function assumes then the user desires calculation of cross-correlations due to these temporal basis functions.}

\item{new_B_spacetime2}{A second new matrix of spatio-temporal basis functions at which cross-calculations are to be calculated.  If this is supplied, then \code{new_B_time} must also be supplied, as the function assumes then the user desires calculation of cross-correlations due to these spatio-temporal basis functions.}
}
\value{
A list with the following components is returned:
\item{correlation: }{A matrix of residual (cross-)correlation values.}
\item{lower: }{A matrix of the lower bound of the uncertainty intervals for the residual correlations.}
\item{upper: }{A matrix of the upper bound of the uncertainty intervals for the residual correlations.}

If \code{se_fit = TRUE}, then a list with the following components (if applicable) is returned:
\item{fit: }{A matrix of predicted values.}
\item{stderr: }{A matrix of standard errors associated with the uncertainty intervals for the predictions.}
\item{lower: }{A matrix of the lower bound of the uncertainty intervals for the predictions.}
\item{upper: }{A matrix of the upper bound of the uncertainty intervals for the predictions.}
Otherwise if \code{se_fit = FALSE}, then a matrix of predicted values is returned.

Other if \code{type = "lpmatrix"}, then a model matrix with the number of rows equal to either the number of rows if \code{object$y} (if \code{newdata} is not supplied) or the number of rows in \code{newdata} when it is supplied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a fitted \code{CBFM} object calculates the residual between-species correlation matrix due to (all) the spatial and/temporal basis functions, along with corresponding uncertainty intervals. Both are constructed via a simulation-based approach. Similar to \code{\link[=predict.CBFM]{predict.CBFM()}}, this correlation matrix can be calculated based on a different sets of basis functions to those used to actually fit the model. Additionally, the user can supplied two sets of spatial and/or temporal basis functions, in which case the function calculates cross-correlations (between and within species) between these two sets of basis functions.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a fitted \code{CBFM} or \code{CBFM_hurdle} object and produces predictions given (potentially) a new set of observational units with their corresponding covariate and basis function functions. Predictions can be accompanied by standard errors, based on the Bayesian covariance matrix of the parameter estimates. As another option, the function can return the model matrix of the covariates constructed (potentially) using at the new set of observational units; in \code{\link[mgcv:predict.gam]{mgcv::predict.gam()}} this is also known as the linear predictor matrix.
}
\details{
This function is somewhat similar to \code{\link[boral:get.residual.cor]{boral::get.residual.cor()}} and \code{\link[gllvm:getResidualCor.gllvm]{gllvm::getResidualCor()}}, in calculating a residual between-species correlation matrix due to (all) the basis functions i.e., residual spatio-temporal correlations, along with corresponding uncertainty intervals. Recall the general form of the mean regression model for the CBFM is given by

\deqn{g(\mu_{ij}) = \eta_{ij} = x_i^\top\beta_j + b_i^\top a_j,}

where \eqn{b_i} denotes a vector of spatial and/or temporal basis functions for unit \eqn{i}, and \eqn{a_j} denotes the corresponding regression coefficients for species \eqn{j}. So for example, users may choose to use the form \eqn{g(\mu_{ij}) = \eta_{ij} = x_i^\top\beta_j + b_{i,space}^\top a_{j,space} + b_{i,time}^\top a_{j,time},} where \eqn{b_i = (b_{i,space}, b_{i,time})} and \eqn{a_j = (a_{j,space}, a_{j,time})}; see also the help file for \code{\link[=CBFM]{CBFM()}} for more information.

One (although by no means the only, see the warning below) approach to thinking about the residual spatio-temporal covariance and hence correlation between two species \eqn{j} and \eqn{j'}, is based on examining the components \eqn{b_i^\top a_j} and \eqn{b_i^\top\beta_{j'}} across the observational units. Both the point estimate and the uncertainty intervals for the residual correlation are constructed by simulation. Specifically, species-specific coefficients corresponding to the basis functions i.e., \eqn{\beta_j}, are sampled from their approximate large sample normal distribution (i.e., basically a Gaussian approximation to the posterior distribution of the parameters; see \code{\link[=CBFM]{CBFM()}} and the section on estimation and inference), which are then used to calculate the correlations. This sampling and calculation is then performed a large number of times (as governed by \code{num_sims}), after which a point estimate of the residual correlations is obtained by taking the sample average, while uncertainty intervals are based on taking sample quantiles.

Note that because this function calculates correlations as based on component of the linear predictor \eqn{b_i^\top a_j}, then it can not be applied to  \code{CBFM_hurdle} object (which by construction contains two linear predictors, so the user has to decide which component of the hurdle model they are interested in).

With the above definition of the residual between-species spatio-temporal correlation, note that the basis functions on which the correlations are constructed need not be the same those used in fitting the original CBFM i.e., the \eqn{b_i}'s can be different to those of \code{object$B}. This is handled via the \code{new_B_space/new_B_time/new_B_spacetime} arguments. Additionally, it is possible to calculate residual within and between species \emph{cross-correlations} across two different sets of spatio-temporal basis functions. That is, residual correlations are calculated between \eqn{b_i^\top a_j} and \eqn{b_{i2}^\top a_{j'}}, where \eqn{b_i} and \eqn{b_{i2}} can be different sets of basis functions. This is handled by supplying both \code{new_B_space/new_B_time/new_B_spacetime} and \code{new_B_space2/new_B_time2/new_B_spacetime2} arguments simultaneously and as appropriate. Cross-correlations may be useful, say, if the two sets of basis functions reflect two different sets of sampling units, and we are interested in how spatially and/or temporally similar (or lack of) the species communities are across these two sets. Another example if is the same set of observational units are visited at two different time points, and we are interested in how similar (or lack thereof) the spatial correlations within and between species are between these two time points (see Thorson et al., 2016; Thorson, 2019; Ovakainen and Abrego, 2020, for similar ideas).

NOTE: A residual cross-correlation matrix is not going to be a standard residual correlation matrix in the sense of having the ones alone the diagonal. This is because even for the same species \eqn{j = j'}, the correlation is not guaranteed to be equal to one as the covariates being considered can be different.

The standard errors produced by \code{predict.CBFM} are based on the Bayesian posterior covariance matrix of the estimated parameters from the fitted \code{CBFM} object, and associated uncertainty intervals are obtained based on the associated large sample normality result of i.e., basically a Gaussian approximation to the posterior distribution of the parameters. This construction is similar to \code{\link[mgcv:predict.gam]{mgcv::predict.gam()}}.

The functions tries to avoid using simulation for constructing the uncertainty intervals for the predictions (to minimize computational burden) However, in some cases it will fallback to doing so when it can not find a simple way to construct these intervals  (blame Francis!) e.g., for the zero-inflated Poisson distribution when \code{type = "response"}.
}
\section{Warning}{
The approach this function takes to construct a residual spatio-temporal correlation matrix between species is \strong{by no means the only approach}. Indeed, with CBFMs the approach taken here to construct residual correlations can be seen as a basis-function analogue of equation 4 in Pollock et al., (2014), as well as Warton et al., (2015) and Hui (2016) among others, for calculating correlations due to measured predictors i.e., except we replace the measure predictors with basis functions.

An alternative approach, and arguably more in line with how residual correlations are calculated in latent variable models (e.g., Warton et al., 2015; Niku et al., 2017; Ovaskainen et al., 2017, Niku et al., 2019), would be to calculate residual spatio-temporal correlations based directly off the community-level covariance matrices \code{Sigma_space/Sigma_time/Sigma_spacetime} and baseline between-species correlation matrices \code{G_space/G_time/G_spacetime} characterizing the random effects covariance of the species-specific coefficients \eqn{(a_1,\ldots,a_m)}. Such an approach is currently not available in CBFM...sorry!
}

\examples{
\donttest{
# Please see examples in the help file for the CBFM and makeahurdle functions 
}

\donttest{
# Please see examples in the help file for the CBFM and makeahurdle functions 
}

}
\references{
Niku, J., Warton, D. I., Hui, F. K., and Taskinen, S. (2017). Generalized linear latent variable models for multivariate count and biomass data in ecology. Journal of Agricultural, Biological and Environmental Statistics, 22, 498-522.

Niku, J., Hui, F. K., Taskinen, S., and Warton, D. I. (2019). gllvm: Fast analysis of multivariate abundance data with generalized linear latent variable models in r. Methods in Ecology and Evolution, 10, 2173-2182.

Ovaskainen, O., Tikhonov, G., Norberg, A., Guillaume Blanchet, F., Duan, L., Dunson, D., and Abrego, N. (2017). How to make more out of community data? A conceptual framework and its implementation as models and software. Ecology letters, 20, 561-576.

Ovaskainen, O., and Abrego, N. (2020). Joint species distribution modelling: with applications in R. Cambridge University Press.

Pollock, L. J., Tingley, R., Morris, W. K., Golding, N., O'Hara, R. B., Parris, K. M., Vesk, P. A., and McCarthy, M. A. (2014). Understanding co‐occurrence by modelling species simultaneously with a Joint Species Distribution Model (JSDM). Methods in Ecology and Evolution, 5, 397-406.

Thorson, J. T. (2019). Guidance for decisions using the Vector Autoregressive Spatio-Temporal (VAST) package in stock, ecosystem, habitat and climate assessments. Fisheries Research, 210, 143-161.

Thorson, J. T., Ianelli, J. N., Larsen, E. A., Ries, L., Scheuerell, M. D., Szuwalski, C., and Zipkin, E. F. (2016). Joint dynamic species distribution models: a tool for community ordination and spatio-temporal monitoring. Global Ecology and Biogeography, 25, 1144-1158.

Warton, D. I., Blanchet, F. G., O'Hara, R. B., Ovaskainen, O., Taskinen, S., Walker, S. C., and Hui, F. K. C. (2015). So many variables: joint modeling in community ecology. Trends in Ecology and Evolution, 30, 766-779.
}
\seealso{
\code{\link[=CBFM]{CBFM()}} for fitting CBFMs, and \code{\link[=corX]{corX()}} for calculating between-species (cross-)correlations due to measured covariates.

\code{\link[=CBFM]{CBFM()}} for fitting CBFMs, \code{\link[=fitted.CBFM]{fitted.CBFM()}} for obtaining fitted values from a CBFM fit, and \code{\link[=residuals.CBFM]{residuals.CBFM()}} for calculating various types of residuals.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak

Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
