% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CBFM.R
\name{print.CBFM}
\alias{print.CBFM}
\title{Print a CBFM object}
\usage{
\method{print}{CBFM}(x, ...)
}
\arguments{
\item{x}{An object of class "CBFM".}

\item{...}{Not used.}
}
\description{
`r lifecycle::badge("stable")

The default print method for a \code{CBFM} object.
}
\details{
Print out details such as the function call, assumed family/response distribution, number of observational units and species, response-environment relationship fitted as given by the formula, which sets of basis functions are used, and the ranks.`
}
\seealso{
\code{\link[=CBFM]{CBFM()}} for fitting CBFMs and \code{\link[=fitted.values.CBFM]{fitted.values.CBFM()}} for calculating fitted values from a CBFM fit.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
