% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CBFM.R
\name{summary.CBFM}
\alias{summary.CBFM}
\title{Summary for a CBFM bit}
\usage{
\method{summary}{CBFM}(
  object,
  coverage = 0.95,
  digits = max(3L, getOption("digits") - 3L),
  ...
)
}
\arguments{
\item{object}{An object of class "CBFM".}

\item{coverage}{The coverage probability of the confidence intervals for the regression coefficients. Defaults to 0.95, which corresponds to 95\% confidence intervals.}

\item{digits}{The number of significant figures to use when printing}

\item{...}{Not used.}
}
\value{
TBC
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Takes a fitted \code{CBFM} object and produces some useful summaries from it.
}
\details{
TBC
}
\examples{
\donttest{
# Please see examples in the help file for the main CBFM function 
}

}
\seealso{
\code{\link[=CBFM]{CBFM()}} for fitting CBFMs and \code{\link[=coef.CBFM]{coef.CBFM()}} for extracting regression coefficients related to the covariates
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
