% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedielogfam.R
\name{tweedielogfam}
\alias{tweedielogfam}
\title{Family object for Tweedie distribution}
\usage{
tweedielogfam()
}
\value{
An object of class "family".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A specified family object for the Tweedie distribution using the log link function and a power mean-variance relationship
}
\details{
This family object was created specifically for fitting CBFMs to non-negative continuous data with a "spike" at zero.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
