% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweedielogfam.R
\name{tweedielogfam}
\alias{tweedielogfam}
\title{Family object for Tweedie distribution}
\usage{
tweedielogfam()
}
\value{
An object of class "family".
}
\description{
\verb{r lifecycle::badge("stable")}

A specified family object for the Tweedie distribution using the log link function and a power mean-variance relationship
}
\details{
This family object was cretated specifically for fitting CBFMs to non-negative continuous data with a "spike" at zero.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
