% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinegativebinomial.R
\name{zinb2}
\alias{zinb2}
\title{Family object for zero-inflated negative binomial distribution}
\usage{
zinb2()
}
\value{
An object of class "family".
}
\description{
\verb{r lifecycle::badge("stable")}

A specified family object for the zero-inflated negative binomial distribution, using the log link function for the negative binomial distribution component.
}
\details{
This family object was created specifically for fitting CBFMs.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
