% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztnegativebinomial.R
\name{ztnb2}
\alias{ztnb2}
\title{Family object for zero-truncated negative binomial distribution}
\usage{
ztnb2()
}
\value{
An object of class "gamlss.family" and "family"
}
\description{
\verb{r lifecycle::badge("stable")}

A specified family object for the zero-truncated negative binomial distribution, using the log link function for the negative binomial distribution component.
}
\details{
This family object was created specifically for fitting CBFMs.

This family object was created specifically for fitting CBFMs. However it relies heavily on \code{\link[gamlss.tr:trun]{gamlss.tr::trun()}}, so big shout outs to the Mikis and the maintainers of that package.
}
\author{
Francis K.C. Hui \href{mailto:fhui28@gmail.com}{fhui28@gmail.com}, Chris Haak
}
